requirements.txt
clip~=1.0
onnx~=1.17.0
torch~=2.6.0
sclblonnx~=0.3.0
onnxruntime~=1.20.1
pillow~=11.1.0
onnxsim~=0.4.36
numpy~=2.2.2